split schedulers
